% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brute_force_knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force search}
\usage{
brute_force_knapsack(x, W)
}
\arguments{
\item{x}{data frame containing variables "v", values, and "w" which is weights of the items.}

\item{W}{Knapsack size}
}
\value{
A list containing the value and the picked elements
}
\description{
Solve the knapsack problem by using brute force search algorithm. This algorithm solves the knapsack problem by going through all possible combinations and return the highest value.
}
