% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greedy_knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy heuristic}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data frame containing variables "v", values, and "w" which is weights of the items.}

\item{W}{Knapsack size}
}
\value{
A list containing the value and the picked elements
}
\description{
Solve the knapsack problem by using a approximating approach.
}
