% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack_dynamic.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Knapsack dynamic}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data frame containing variables "v", values, and "w" which is weights of the items.}

\item{W}{Knapsack size}
}
\value{
A list containing the value and the picked elements
}
\description{
Solve the knapsack problem by dynamic programming. This algorithm solves the knapsack problem by iterating over all possible values of w.
}
